Refactoring BigGAN part 1
Split the BigGAN model, individuals, image processing, and interpolation into separate classes.
Additionally started on a simple logger for maintaining the state of the evolutionary proces.
Furthermore started with some initial tests on the individual and the logger.
All source code is now maintained within the src folder, tests in test, log files in logs, and images should be placed in imgs.