Primitive Setup for candidate selection
Started from scratch on a new GUI to see if the crashes would stop. The new GUI (evolution_GUI) is still in a primitive stage, but is more reliable in updating and displaying parents, and candidates.
Warning, found staleness in the GAN process due to finally having the sequential process of candidate selection.